+++ src/IB.CSharpApiClient/Messages/OrderStatusMessage.cs
@@ -3,8 +3,8 @@
     public class OrderStatusMessage : OrderMessage
     {
         public string Status { get; private set; }
-        public double Filled { get; private set; }
-        public double Remaining { get; private set; }
+        public decimal Filled { get; private set; }
+        public decimal Remaining { get; private set; }
         public double AvgFillPrice { get; private set; }
         public int PermId { get; private set; }
         public int ParentId { get; private set; }
@@ -13,7 +13,7 @@
         public string WhyHeld { get; private set; }
         public double MktCapPrice { get; private set; }
 
-        public OrderStatusMessage(int orderId, string status, double filled, double remaining, double avgFillPrice,
+        public OrderStatusMessage(int orderId, string status, decimal filled, decimal remaining, double avgFillPrice,
            int permId, int parentId, double lastFillPrice, int clientId, string whyHeld, double mktCapPrice)
         {
             OrderId = orderId;
